Bug bashes fixes
