+++ packages/reown_appkit/lib/modal/pages/public/appkit_modal_main_wallets_page.dart
@@ -187,6 +187,7 @@ class _AppKitModalMainWalletsPageState
                                 const ReownAppKitModalQRCodePage(),
                                 event: SelectWalletEvent(
                                   name: 'WalletConnect',
+                                  explorerId: '',
                                   platform: AnalyticsPlatform.qrcode,
                                 ),
                               );
