+++ packages/reown_appkit/lib/modal/pages/select_token_page.dart
@@ -46,7 +46,7 @@ class _SelectTokenPageState extends State<SelectTokenPage> {
         } else {
           _tokens = await _blockchainService.getBalance(
             address: address,
-            caip2chain: '$namespace:$chainId',
+            caip2Chain: '$namespace:$chainId',
           );
         }
         setState(() {});
