+++ packages/reown_appkit/lib/modal/services/uri_service/url_utils.dart
@@ -96,7 +96,7 @@ class UriService extends IUriService {
       }
       return true;
     } catch (e) {
-      rethrow;
+      throw CanNotLaunchUrl();
     }
   }
