+++ packages/reown_appkit/lib/modal/widgets/miscellaneous/all_wallets_header.dart
@@ -42,6 +42,7 @@ class AllWalletsHeader extends StatelessWidget {
                 const ReownAppKitModalQRCodePage(),
                 event: SelectWalletEvent(
                   name: 'WalletConnect',
+                  explorerId: '',
                   platform: AnalyticsPlatform.qrcode,
                 ),
               );
