+++ packages/reown_appkit/pubspec.yaml
@@ -35,10 +35,16 @@ dependencies:
   # reown_sign:
   #   path: ../reown_sign/
   shimmer: ^3.0.0
+  solana_web3: ^0.1.3
   uuid: ^4.5.1
   webview_flutter: 4.10.0
   webview_flutter_android: 4.2.0
   webview_flutter_wkwebview: 3.17.0
+
+dependency_overrides:
+  # because of solana_web3
+  pinenacl: ^0.6.0
+  web_socket_channel: ^3.0.1
 
 dev_dependencies:
   build_runner: ^2.4.13
